requirements.txt
absl-py==0.13.0
attrs==21.2.0
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
cvzone==1.5.1
cycler==0.10.0
dlib==19.22.1
face-recognition==1.3.0
face-recognition-models==0.3.0
idna==3.2
importlib-metadata==4.8.1
keyboard==0.13.5
kiwisolver==1.3.2
matplotlib==3.4.3
mediapipe==0.8.7.2
numpy==1.21.2
opencv-contrib-python==4.5.3.56
opencv-python==4.5.3.56
Pillow==8.3.2
playsound==1.2.2
protobuf==3.17.3
pyparsing==2.4.7
python-dateutil==2.8.2
requests==2.26.0
six==1.16.0
typing-extensions==3.10.0.2
urllib3==1.26.6
zipp==3.5.0
